Modified hotel pages
Modifications were made to the hotel pages and the home page.

NAV was modified
Implemented a new design in the hotel selection.

Modified hotel pages

Modifications were made to the hotel pages and the home page.

NAV was modified
Implemented a new design in the hotel selection.

In the hotel selection page a template design was used as a template for all hotel pages.